compiled for android

+++ dlib/global_optimization/global_function_search.cpp
@@ -302,7 +302,7 @@ namespace dlib
             {
                 temp(i) = rnd.get_double_in_range(lower(i), upper(i));
                 if (is_integer_variable[i])
-                    temp(i) = std::round(temp(i));
+                    temp(i) = roundl(temp(i));
             }
             return temp;
         }
@@ -403,9 +403,9 @@ namespace dlib
         {
             if (is_integer_variable[i])
             {
-                DLIB_CASSERT(std::round(lower(i)) == lower(i), "If you say a variable is an integer variable then it must have an integer lower bound. \n"
+                DLIB_CASSERT(roundl(lower(i)) == lower(i), "If you say a variable is an integer variable then it must have an integer lower bound. \n"
                     << "lower[i] = " << lower(i));
-                DLIB_CASSERT(std::round(upper(i)) == upper(i), "If you say a variable is an integer variable then it must have an integer upper bound. \n"
+                DLIB_CASSERT(roundl(upper(i)) == upper(i), "If you say a variable is an integer variable then it must have an integer upper bound. \n"
                     << "upper[i] = " << upper(i));
             }
         }
@@ -714,7 +714,7 @@ namespace dlib
                 for (long i = 0; i < new_req.x.size(); ++i)
                 {
                     if (info->spec.is_integer_variable[i])
-                        new_req.x(i) = std::round(new_req.x(i));
+                        new_req.x(i) = roundl(new_req.x(i));
                 }
                 info->outstanding_evals.emplace_back(new_req);
                 return function_evaluation_request(new_req,info);
